+++ v24/api/keyfactor/v1/model_certificate_authorities_certificate_authority_response.go
@@ -73,6 +73,11 @@ type CertificateAuthoritiesCertificateAuthorityResponse struct {
 	Audience                      NullableString                                                                  `json:"Audience,omitempty"`
 	ClientSecret                  *CSSCMSDataModelModelsKeyfactorAPISecret                                        `json:"ClientSecret,omitempty"`
 	LastScan                      NullableString                                                                  `json:"LastScan,omitempty"`
+	UseForEnrollment              *bool                                                                           `json:"UseForEnrollment,omitempty"`
+	CertificateCleanupEnabled     NullableBool                                                                    `json:"CertificateCleanupEnabled,omitempty"`
+	DeleteWithArchivedKey         NullableBool                                                                    `json:"DeleteWithArchivedKey,omitempty"`
+	TimeAfterExpiration           NullableInt32                                                                   `json:"TimeAfterExpiration,omitempty"`
+	TimeAfterExpirationUnits      *CSSCMSDataModelEnumsCertificateCleanupTimeUnits                                `json:"TimeAfterExpirationUnits,omitempty"`
 }
 
 // NewCertificateAuthoritiesCertificateAuthorityResponse instantiates a new CertificateAuthoritiesCertificateAuthorityResponse object
@@ -1721,6 +1726,199 @@ func (o *CertificateAuthoritiesCertificateAuthorityResponse) UnsetLastScan() {
 	o.LastScan.Unset()
 }
 
+// GetUseForEnrollment returns the UseForEnrollment field value if set, zero value otherwise.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetUseForEnrollment() bool {
+	if o == nil || isNil(o.UseForEnrollment) {
+		var ret bool
+		return ret
+	}
+	return *o.UseForEnrollment
+}
+
+// GetUseForEnrollmentOk returns a tuple with the UseForEnrollment field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetUseForEnrollmentOk() (*bool, bool) {
+	if o == nil || isNil(o.UseForEnrollment) {
+		return nil, false
+	}
+	return o.UseForEnrollment, true
+}
+
+// HasUseForEnrollment returns a boolean if a field has been set.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) HasUseForEnrollment() bool {
+	if o != nil && !isNil(o.UseForEnrollment) {
+		return true
+	}
+
+	return false
+}
+
+// SetUseForEnrollment gets a reference to the given bool and assigns it to the UseForEnrollment field.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) SetUseForEnrollment(v bool) {
+	o.UseForEnrollment = &v
+}
+
+// GetCertificateCleanupEnabled returns the CertificateCleanupEnabled field value if set, zero value otherwise.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetCertificateCleanupEnabled() bool {
+	if o == nil || !o.CertificateCleanupEnabled.IsSet() || o.CertificateCleanupEnabled.Get() == nil {
+		var ret bool
+		return ret
+	}
+	return *o.CertificateCleanupEnabled.Get()
+}
+
+// GetCertificateCleanupEnabledOk returns a tuple with the CertificateCleanupEnabled field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+// NOTE: If the value is an explicit nil, `nil, true` will be returned
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetCertificateCleanupEnabledOk() (*bool, bool) {
+	if o == nil {
+		return nil, false
+	}
+	return o.CertificateCleanupEnabled.Get(), o.CertificateCleanupEnabled.IsSet()
+}
+
+// HasCertificateCleanupEnabled returns a boolean if a field has been set.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) HasCertificateCleanupEnabled() bool {
+	if o != nil && o.CertificateCleanupEnabled.IsSet() {
+		return true
+	}
+
+	return false
+}
+
+// SetCertificateCleanupEnabled gets a reference to the given NullableBool and assigns it to the CertificateCleanupEnabled field.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) SetCertificateCleanupEnabled(v bool) {
+	o.CertificateCleanupEnabled.Set(&v)
+}
+
+// SetCertificateCleanupEnabledNil sets the value for CertificateCleanupEnabled to be an explicit nil
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) SetCertificateCleanupEnabledNil() {
+	o.CertificateCleanupEnabled.Set(nil)
+}
+
+// UnsetCertificateCleanupEnabled ensures that no value is present for CertificateCleanupEnabled, not even an explicit nil
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) UnsetCertificateCleanupEnabled() {
+	o.CertificateCleanupEnabled.Unset()
+}
+
+// GetDeleteWithArchivedKey returns the DeleteWithArchivedKey field value if set, zero value otherwise.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetDeleteWithArchivedKey() bool {
+	if o == nil || !o.DeleteWithArchivedKey.IsSet() || o.DeleteWithArchivedKey.Get() == nil {
+		var ret bool
+		return ret
+	}
+	return *o.DeleteWithArchivedKey.Get()
+}
+
+// GetDeleteWithArchivedKeyOk returns a tuple with the DeleteWithArchivedKey field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+// NOTE: If the value is an explicit nil, `nil, true` will be returned
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetDeleteWithArchivedKeyOk() (*bool, bool) {
+	if o == nil {
+		return nil, false
+	}
+	return o.DeleteWithArchivedKey.Get(), o.DeleteWithArchivedKey.IsSet()
+}
+
+// HasDeleteWithArchivedKey returns a boolean if a field has been set.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) HasDeleteWithArchivedKey() bool {
+	if o != nil && o.DeleteWithArchivedKey.IsSet() {
+		return true
+	}
+
+	return false
+}
+
+// SetDeleteWithArchivedKey gets a reference to the given NullableBool and assigns it to the DeleteWithArchivedKey field.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) SetDeleteWithArchivedKey(v bool) {
+	o.DeleteWithArchivedKey.Set(&v)
+}
+
+// SetDeleteWithArchivedKeyNil sets the value for DeleteWithArchivedKey to be an explicit nil
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) SetDeleteWithArchivedKeyNil() {
+	o.DeleteWithArchivedKey.Set(nil)
+}
+
+// UnsetDeleteWithArchivedKey ensures that no value is present for DeleteWithArchivedKey, not even an explicit nil
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) UnsetDeleteWithArchivedKey() {
+	o.DeleteWithArchivedKey.Unset()
+}
+
+// GetTimeAfterExpiration returns the TimeAfterExpiration field value if set, zero value otherwise (both if not set or set to explicit null).
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetTimeAfterExpiration() int32 {
+	if o == nil || isNil(o.TimeAfterExpiration.Get()) {
+		var ret int32
+		return ret
+	}
+	return *o.TimeAfterExpiration.Get()
+}
+
+// GetTimeAfterExpirationOk returns a tuple with the TimeAfterExpiration field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+// NOTE: If the value is an explicit nil, `nil, true` will be returned
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetTimeAfterExpirationOk() (*int32, bool) {
+	if o == nil {
+		return nil, false
+	}
+	return o.TimeAfterExpiration.Get(), o.TimeAfterExpiration.IsSet()
+}
+
+// HasTimeAfterExpiration returns a boolean if a field has been set.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) HasTimeAfterExpiration() bool {
+	if o != nil && o.TimeAfterExpiration.IsSet() {
+		return true
+	}
+
+	return false
+}
+
+// SetTimeAfterExpiration gets a reference to the given NullableInt32 and assigns it to the TimeAfterExpiration field.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) SetTimeAfterExpiration(v int32) {
+	o.TimeAfterExpiration.Set(&v)
+}
+
+// SetTimeAfterExpirationNil sets the value for TimeAfterExpiration to be an explicit nil
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) SetTimeAfterExpirationNil() {
+	o.TimeAfterExpiration.Set(nil)
+}
+
+// UnsetTimeAfterExpiration ensures that no value is present for TimeAfterExpiration, not even an explicit nil
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) UnsetTimeAfterExpiration() {
+	o.TimeAfterExpiration.Unset()
+}
+
+// GetTimeAfterExpirationUnits returns the TimeAfterExpirationUnits field value if set, zero value otherwise.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetTimeAfterExpirationUnits() CSSCMSDataModelEnumsCertificateCleanupTimeUnits {
+	if o == nil || isNil(o.TimeAfterExpirationUnits) {
+		var ret CSSCMSDataModelEnumsCertificateCleanupTimeUnits
+		return ret
+	}
+	return *o.TimeAfterExpirationUnits
+}
+
+// GetTimeAfterExpirationUnitsOk returns a tuple with the TimeAfterExpirationUnits field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) GetTimeAfterExpirationUnitsOk() (*CSSCMSDataModelEnumsCertificateCleanupTimeUnits, bool) {
+	if o == nil || isNil(o.TimeAfterExpirationUnits) {
+		return nil, false
+	}
+	return o.TimeAfterExpirationUnits, true
+}
+
+// HasTimeAfterExpirationUnits returns a boolean if a field has been set.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) HasTimeAfterExpirationUnits() bool {
+	if o != nil && !isNil(o.TimeAfterExpirationUnits) {
+		return true
+	}
+
+	return false
+}
+
+// SetTimeAfterExpirationUnits gets a reference to the given CSSCMSDataModelEnumsCertificateCleanupTimeUnits and assigns it to the TimeAfterExpirationUnits field.
+func (o *CertificateAuthoritiesCertificateAuthorityResponse) SetTimeAfterExpirationUnits(v CSSCMSDataModelEnumsCertificateCleanupTimeUnits) {
+	o.TimeAfterExpirationUnits = &v
+}
+
 func (o CertificateAuthoritiesCertificateAuthorityResponse) MarshalJSON() ([]byte, error) {
 	toSerialize, err := o.ToMap()
 	if err != nil {
@@ -1862,6 +2060,21 @@ func (o CertificateAuthoritiesCertificateAuthorityResponse) ToMap() (map[string]
 	}
 	if o.LastScan.IsSet() {
 		toSerialize["LastScan"] = o.LastScan.Get()
+	}
+	if !isNil(o.UseForEnrollment) {
+		toSerialize["UseForEnrollment"] = o.UseForEnrollment
+	}
+	if o.CertificateCleanupEnabled.IsSet() {
+		toSerialize["CertificateCleanupEnabled"] = o.CertificateCleanupEnabled.Get()
+	}
+	if o.DeleteWithArchivedKey.IsSet() {
+		toSerialize["DeleteWithArchivedKey"] = o.DeleteWithArchivedKey.Get()
+	}
+	if o.TimeAfterExpiration.IsSet() {
+		toSerialize["TimeAfterExpiration"] = o.TimeAfterExpiration.Get()
+	}
+	if !isNil(o.TimeAfterExpirationUnits) {
+		toSerialize["TimeAfterExpirationUnits"] = o.TimeAfterExpirationUnits
 	}
 	return toSerialize, nil
 }